% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGeneData.R
\name{extractGeneData}
\alias{extractGeneData}
\title{Extract Gene Data}
\usage{
extractGeneData(
  Query_Object,
  Gene.list = genes$gene_id,
  sample.type = "Primary solid Tumor",
  columns.to.check = names(clinical)
)
}
\description{
Extract Gene Data
}
