% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDCfilterByDC.R
\name{GDCfilterByDC}
\alias{GDCfilterByDC}
\title{GDC filter based on data category}
\usage{
GDCfilterByDC(data.category)
}
\arguments{
\item{data.category}{: The data category entered by the user.}
}
\value{
list of each project that has the desired data category entered by the user.
}
\description{
GDC filter based on data category
}
\details{
It enters to GDC project , retrieve the size of each project. Then, it shows the data category of each project. Finally, it  filters the data based on the desired data category.
}
