% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDCproject.R
\name{GDCproject}
\alias{GDCproject}
\title{Function retrieves All GDC project summary}
\usage{
GDCproject()
}
\value{
environment with the summary of each project in the GDC data base.
}
\description{
Function retrieves All GDC project summary
}
\details{
GDC project function return list of project IDs including summary of each project. This summary includes data category, case count and file count.
}
\examples{
GDCproject()
}
