% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneDataExtract.R
\name{GeneDataExtract}
\alias{GeneDataExtract}
\title{Create a data frame for Survival Analysis}
\usage{
GeneDataExtract(
  object,
  Gene.list = Gene.list,
  limma_res,
  method = "KM_",
  plot = TRUE,
  adj.P.value_Cutoff = 0.05,
  logFC_Cutoff = 1,
  columns.to.check = c()
)
}
\arguments{
\item{object}{a SummerizedExperimental object.}

\item{limma_res}{Object by limma package}

\item{method}{the method to detect UP and DOWN expression level of certain gene according to soecific cut off, method = ("quantiles","median","KM_","CoxPH_")}

\item{columns.to.check}{character or vector of characters to be extracted from samples data.}
}
\value{
a data frame contains the gene expression row values, vital status, days to last follow up, normalized gene expression values and the level of expression as well as other columns specified.
}
\description{
Create a data frame for Survival Analysis
}
\examples{
GeneDataExtract(object = TCGAsurvival::Brain_TCGA_GBM_Object , Gene.list = c("FBP1" , "MAPK1") , limma_res = TCGAsurvival::Brain_limma_res , method = "KM_")
}
