% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfIntegration.R
\name{integrate_files}
\alias{integrate_files}
\title{Integrate Data Frames}
\usage{
integrate_files(
  to_dataframe,
  from_dataframe,
  key_to,
  key_from,
  columns.to.check = key_from,
  empty.values = "NULL"
)
}
\description{
Integrate Data Frames
}
