% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplot}
\alias{PCAplot}
\title{PCA Plotting Funcion}
\usage{
PCAplot(limma_res, condition_variable = "definition", Pos = "topleft")
}
\arguments{
\item{limma_res:}{an object, (output of TCGAnormalize_limma(...))}

\item{condition_variable:}{"definition" column in the limma_res object by default.}

\item{Pos:}{Position of the legend in the plot (default #topleft) #buttomleft #topright ..}
}
\value{
: a data frame with all principal components.
}
\description{
PCA Plotting Funcion
}
\examples{
PCAplot(limma_res = TCGAsurvival::Brain_limma_res)

}
