% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj_inPS.R
\name{Projects_in_Primary_sites}
\alias{Projects_in_Primary_sites}
\title{Each Primary Tumor Site & The Corresponding Projects}
\usage{
Projects_in_Primary_sites()
}
\value{
: Dict. or Environment, each key represents a primary tumor site and the values are the coressponding projects
}
\description{
Each Primary Tumor Site & The Corresponding Projects
}
\examples{
PS <- Projects_in_Primary_sites()
}
