% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSelect.R
\name{SampleSeclect}
\alias{SampleSeclect}
\title{Sample Select}
\usage{
SampleSeclect(query, sample.type = "Primary Tumor", nT, nN, S = S)
}
\arguments{
\item{query:}{an object (output of TCGAtumorSiteSelect(...))}

\item{sample.type:}{Type of tumor}

\item{nT:}{number of Tumor samples}

\item{nN:}{number of Normal samples}

\item{S:}{set.seed function argument}
}
\value{
list of samples to be downloaded from GDC database
}
\description{
Sample Select
}
