% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_pipeline.R
\name{TCGAnormalize_limma}
\alias{TCGAnormalize_limma}
\title{Limma Function For Normalization}
\usage{
TCGAnormalize_limma(
  object,
  condition_variable = "definition",
  reference_group = "Solid Tissue Normal",
  Norm.Method = "TMM",
  plot = TRUE,
  rm.zeros = TRUE
)
}
\arguments{
\item{object}{Summerizedexperimental object}

\item{condition_variable}{definition feature exported from colData from TCGA object}

\item{reference_group}{To decide which one of the values of your conditional variable is going to be the reference group}

\item{Norm.Method}{Normalization method , method=c("TMM","RLE","upperquartile","none")}

\item{plot}{a logical variable, if TRUE plot the mean variance scatter plot}

\item{rm.zeros}{a logical variable, if TRUE removes genes that have zero expression accross all samples before normalization}
}
\value{
a list of voom: A complex object   contains the TMM+voom normalized data;
        eBayes: A complex object  contains the the fitted model plus a number of statistics related to each of the probes
        topTable: A simple table contains  the  differentially expressed genes sorted by p-value
}
\description{
Limma Function For Normalization
}
\examples{
TCGAnormalize_limma(TCGAsurvival::Brain_TCGA_GBM_Object , condition_variable= 'definition',reference_group="Solid Tissue Normal")
}
