% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAsurvKM.R
\name{TCGAsurv}
\alias{TCGAsurv}
\title{Fit a Survival Model For a List of Genes}
\usage{
TCGAsurv(Normalized_df, limma_res, Gene.list, method = "KM", plot = TRUE)
}
\arguments{
\item{Normalized_df:}{a data frame contains the normalized gene (or list of genes) expression values, and its corresponding level of expression.}

\item{plot:}{logical. if TRUE the KM plot will be displayed.}
}
\value{
For method "KM" returns fitted model including P Value of the model.
}
\description{
Fit a Survival Model For a List of Genes
}
