% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{volcPlot}
\alias{volcPlot}
\title{Plotting a Volcano PLot Display DEGs}
\usage{
volcPlot(limma_res, logFC_Cutoff = 2, adj.p.Value_Cutoff = 0.05)
}
\arguments{
\item{limma_res}{an object (output of TCGAnormalize_limma()). the object contains the normalized genes' expression values and the fitted model, as well as the genes' names.}

\item{logFC_Cutoff}{Selected cut off of the log Fold-Change}

\item{adj.p.Value_Cutoff}{adjusted P value cutoff to evaluate the differential gene expression significance.}
}
\value{
displays a Volcano plot with Up and Down differentially expressed Genes
}
\description{
Plotting a Volcano PLot Display DEGs
}
\examples{
volcPlot(limma_res = TCGAsurvival::Brain_limma_res, logFC_Cutoff = 2,adj.p.Value_Cutoff = 0.05)
}
